added model test, renamed progressiveganv2 to progressivegan
