+++ nobrainer/models/progressivegan.py
@@ -5,6 +5,8 @@ import math
 import tensorflow as tf
 from tensorflow.keras import layers, models
 
+import nobrainer
+from nobrainer.volume import adjust_dynamic_range as _adjust_dynamic_range 
 
 def progressivegan(
     latent_size,
@@ -42,8 +44,7 @@ def progressivegan(
         num_channels=num_channels, 
         dimensionality=dimensionality,
         fmap_base=g_fmap_base,
-        fmap_max=g_fmap_max,
-        name=name+'_generator'
+        fmap_max=g_fmap_max
     )
 
     discriminator = Discriminator(
@@ -51,18 +52,15 @@ def progressivegan(
         num_channels=num_channels, 
         dimensionality=dimensionality,
         fmap_base=d_fmap_base,
-        fmap_max=d_fmap_max,
-        name=name+'_discriminator'
+        fmap_max=d_fmap_max
     )
 
     return generator, discriminator
 
 
-class Generator:
-    '''
-    Progressive Generator
-    '''
-    def __init__(self, latent_size, label_size=0, num_channels=1, fmap_base=8192, fmap_max=256, dimensionality=3, name='generator'):
+class Generator(tf.keras.Model):
+
+    def __init__(self, latent_size, label_size=0, num_channels=1, fmap_base=8192, fmap_max=256, dimensionality=3):
         super(Generator, self).__init__()
 
         self.latent_size = latent_size
@@ -72,16 +70,22 @@ class Generator:
         self.fmap_max = fmap_max
         self.num_channels = num_channels
         self.dimensionality = dimensionality
-        self.name = name
-
-        self.growing_generator = self._make_generator_base()
-        self.train_generator = self.growing_generator
-
-        self.current_resolution = 1
 
         self.Conv =  getattr(layers, 'Conv{}D'.format(self.dimensionality))
         self.ConvTranspose = getattr(layers, 'Conv{}DTranspose'.format(self.dimensionality))
         self.Upsampling = getattr(layers, 'UpSampling{}D'.format(self.dimensionality))
+
+        self.current_resolution = 2
+        self.highest_resolution_block = self._make_generator_block(
+            self._nf(self.current_resolution), name='g_block_{}'.format(self.current_resolution))
+
+        self.resolution_blocks = []
+
+        self.base_dense = tf.keras.layers.Dense(units=self._nf(1)*(2**self.dimensionality))
+        self.HeadConv1 = self.Conv(filters=self.num_channels, kernel_size=1)
+        self.HeadConv2 = self.Conv(filters=self.num_channels, kernel_size=1)
+
+        self.build([(None, latent_size), (1,)])
 
     def _pixel_norm(self, epsilon=1e-8):
         '''
@@ -102,20 +106,18 @@ class Generator:
         '''  
         return min(int(self.fmap_base / (2.0 ** (stage))), self.fmap_max)
 
-    def _make_generator_base(self):
+    def generator_base(self, latents):
         '''
         Creates the base of the generator with inputs of size latent_size+label_size
         and the alpha value to determine level of interpolation
         '''
-        latents = layers.Input(shape=[self.latent_size+self.label_size], name='latents')
-        alpha = layers.Input(shape=[], dtype=tf.float32, name='g_alpha')
 
         # Latents stage
         x = self._pixel_norm()(latents)
-        x = layers.Dense(self._nf(1)*(2**self.dimensionality))(x)
+        x = self.base_dense(x)
         x = layers.Reshape([2]*self.dimensionality+[self._nf(1)])(x)
 
-        return models.Model(inputs=[latents, alpha], outputs=[x], name='generator_base')
+        return x
 
     def _make_generator_block(self, nf, kernel_size=4, name=''):
         '''
@@ -126,57 +128,70 @@ class Generator:
 
         # block_layers.append(self.ConvTranspose(nf, kernel_size=3, strides=2, padding='same'))
         block_layers.append(self.Upsampling())
-        #block_layers.append(self.Conv(nf, kernel_size=kernel_size, strides=1, padding='same'))
-        #block_layers.append(layers.Activation(tf.nn.leaky_relu))
-        #block_layers.append(self._pixel_norm())
+        # block_layers.append(self.Conv(filters=nf, kernel_size=kernel_size, strides=1, padding='same'))
+        # block_layers.append(layers.Activation(tf.nn.leaky_relu))
+        # block_layers.append(self._pixel_norm())
 
-        # block_layers.append(self.Conv(nf, kernel_size=3, strides=1, padding='same'))
-        block_layers.append(self.Conv(nf, kernel_size=kernel_size, strides=1, padding='same'))
+        block_layers.append(self.Conv(filters=nf, kernel_size=kernel_size, strides=1, padding='same'))
         block_layers.append(layers.Activation(tf.nn.leaky_relu))
         block_layers.append(self._pixel_norm())
 
         return models.Sequential(block_layers, name=name)
 
-    def add_resolution(self):
-        '''
-        Adds a resolution step to the trainable generator by creating a generator block and appending
-        on top of the existing generator to increase the resolution of the images generated by a factor of 2.
-        The new layers are faded in and controlled by the alpha parameter
-        '''
+    def generator_head(self, x, y, alpha):
 
-        self.current_resolution += 1
+        x = self.Upsampling()(x)
+        x = self.HeadConv1(x)
 
-        # Residual from input
-        to_rgb_1 = self.Upsampling()(self.growing_generator.output)
-        to_rgb_1 = self.Conv(self.num_channels, kernel_size=1)(to_rgb_1)
-       
-        # Growing generator
-        g_block = self._make_generator_block(self._nf(self.current_resolution), name='g_block_{}'.format(self.current_resolution))
-        g_block_output = g_block(self.growing_generator.output)
-        to_rgb_2 = self.Conv(self.num_channels, kernel_size=1)(g_block_output)
+        y = self.HeadConv2(y)
 
-        output = self._weighted_sum()([to_rgb_1, to_rgb_2, self.growing_generator.input[1]])
+        output = self._weighted_sum()([x, y, alpha])
         output = layers.Activation('tanh')(output)
 
-        self.growing_generator = models.Model(inputs=self.growing_generator.input, outputs=g_block_output)
-        self.train_generator = models.Model(inputs=self.growing_generator.input, outputs=[output], name=self.name)
+        return output
 
-    def get_current_resolution(self):
-        return self.current_resolution
+    def add_resolution(self):
 
-    def get_trainable_generator(self):
-        return self.train_generator
+        self.current_resolution += 1
+        self.resolution_blocks.append(self.highest_resolution_block)
+        self.highest_resolution_block = self._make_generator_block(
+            self._nf(self.current_resolution), name='g_block_{}'.format(self.current_resolution))
 
-    def get_inference_generator(self):
-        raise NotImplementedError
+        self.HeadConv1 = self.Conv(filters=self.num_channels, kernel_size=1)
+        self.HeadConv2 = self.Conv(filters=self.num_channels, kernel_size=1)
+        
+        self.build([(None, self.latent_size), (1,)])
 
+    def call(self, inputs):
+        latents, alpha = inputs
+        x = self.generator_base(latents)
 
-class Discriminator:
+        for g_block in self.resolution_blocks:
+            x = g_block(x)
+
+        y = self.highest_resolution_block(x)
+        return self.generator_head(x, y, alpha)
+
+    def generate(self, latents):
+        alpha = tf.constant([1.0], tf.float32)
+        image = self.call([latents, alpha])
+        image = _adjust_dynamic_range(image, [-1,1], [0,255])
+        return {'generated': image} 
+
+    def save(self, filepath, **kwargs):
+         # force retrace of tf.function to track new and deleted variables
+        super().save(
+            filepath, 
+            signatures=tf.function(self.generate, 
+                input_signature=(tf.TensorSpec(shape=[None, self.latent_size], dtype=tf.float32),)), 
+            **kwargs)
+
+class Discriminator(tf.keras.Model):
     '''
     Progressive Discriminator
     '''
 
-    def __init__(self, label_size=0, num_channels=1, fmap_base=8192, fmap_max=512, dimensionality=3, name='discriminator'):
+    def __init__(self, label_size=0, num_channels=1, fmap_base=8192, fmap_max=512, dimensionality=3):
         super(Discriminator, self).__init__()
 
         self.label_size = label_size
@@ -185,15 +200,25 @@ class Discriminator:
         self.fmap_max = fmap_max
         self.num_channels = num_channels
         self.dimensionality = dimensionality
-        self.name = name
 
-        self.growing_discriminator = self._make_discriminator_base()
-        self.train_discriminator = self.growing_discriminator
-
-        self.current_resolution = 1
+        self.current_resolution = 2
 
         self.Conv =  getattr(layers, 'Conv{}D'.format(self.dimensionality))
         self.AveragePooling =  getattr(layers, 'AveragePooling{}D'.format(self.dimensionality))
+
+        self.highest_resolution_block = self._make_discriminator_block(
+            self._nf(self.current_resolution-1), name='d_block_{}'.format(self.current_resolution))
+        self.resolution_blocks = []
+
+        self.BaseConv = self.Conv(filters=self._nf(self.current_resolution-1), kernel_size=1, padding='same')
+        self.FadeConv = self.Conv(filters=self._nf(self.current_resolution), kernel_size=1, padding='same')
+        self.HeadDense1 = tf.keras.layers.Dense(units=self._nf(1))
+        self.HeadDense2 = tf.keras.layers.Dense(units=1+self.label_size)
+
+        images_shape = (None,)+(int(2.0**self.current_resolution),)*self.dimensionality + (self.num_channels,)
+        alpha_shape = (1,)
+
+        self.build([images_shape, alpha_shape])
 
     def _weighted_sum(self):
         '''
@@ -207,22 +232,14 @@ class Discriminator:
         '''  
         return min(int(self.fmap_base / (2.0 ** (stage))), self.fmap_max)
 
-    def _make_discriminator_base(self):
-        '''
-        Creates the end of the disciminator with inputs of image dimensions and the 
-        alpha value to determine level of interpolation. The output node gives
-        the prediction of fakeness of the input image.
-        '''
+    def discriminator_base(self, x, y, alpha):
 
-        inputs = layers.Input(shape=(2,)*self.dimensionality + (self._nf(1),), name='dummy')
+        x = self.AveragePooling()(x)
+        x = self.BaseConv(x)
 
-        x = layers.Flatten()(inputs)
-        x = layers.Dense(self._nf(1))(x)
-        x = layers.Activation(tf.nn.leaky_relu)(x)
+        lerp_input = self._weighted_sum()([x, y, alpha])
 
-        output = layers.Dense(1+self.label_size)(x)
-
-        return models.Model(inputs=[inputs], outputs=[output])
+        return lerp_input
 
     def _make_discriminator_block(self, nf, kernel_size=4, name=''):
         '''
@@ -231,13 +248,30 @@ class Discriminator:
 
         block_layers = []
 
-        #block_layers.append(self.Conv(nf, kernel_size=kernel_size, strides=1, padding='same'))
-        #block_layers.append(layers.Activation(tf.nn.leaky_relu))
+        # block_layers.append(self.Conv(filters=nf, kernel_size=kernel_size, strides=1, padding='same'))
+        # block_layers.append(layers.Activation(tf.nn.leaky_relu))
 
-        block_layers.append(self.Conv(nf, kernel_size=kernel_size, strides=2, padding='same'))
+        block_layers.append(self.Conv(filters=nf, kernel_size=kernel_size, strides=2, padding='same'))
         block_layers.append(layers.Activation(tf.nn.leaky_relu))
 
         return models.Sequential(block_layers, name=name)
+
+    def discriminator_head(self, x):
+        '''
+        Creates the end of the disciminator with inputs of image dimensions and the 
+        alpha value to determine level of interpolation. The output node gives
+        the prediction of fakeness of the input image.
+        '''
+        x = layers.Flatten()(x)
+        x = self.HeadDense1(x)
+        x = layers.Activation(tf.nn.leaky_relu)(x)
+
+        output = self.HeadDense2(x)
+
+        score_output = layers.Lambda(lambda x: x[...,0])(output)
+        label_output = layers.Lambda(lambda x: x[...,1:])(output)
+
+        return score_output, label_output
 
     def add_resolution(self):
         '''
@@ -248,33 +282,28 @@ class Discriminator:
 
         self.current_resolution += 1
 
-        inputs = layers.Input(shape=(int(2.0**self.current_resolution),)*self.dimensionality + (self.num_channels,), name='image')
-        alpha = layers.Input(shape=[], name='d_alpha')
+        self.resolution_blocks.append(self.highest_resolution_block)
+        self.highest_resolution_block = self._make_discriminator_block(
+            self._nf(self.current_resolution-1), name='d_block_{}'.format(self.current_resolution))
 
-        # Residual from input
-        from_rgb_1 = self.AveragePooling()(inputs)
-        from_rgb_1 = self.Conv(self._nf(self.current_resolution-1), kernel_size=1, padding='same', name='from_rgb_1')(from_rgb_1)
+        self.BaseConv = self.Conv(filters=self._nf(self.current_resolution-1), kernel_size=1, padding='same')
+        self.FadeConv = self.Conv(filters=self._nf(self.current_resolution), kernel_size=1, padding='same')
 
-        # Growing discriminator
-        d_block = self._make_discriminator_block(self._nf(self.current_resolution-1), name='d_block_{}'.format(self.current_resolution))
-        from_rgb_2 = self.Conv(self._nf(self.current_resolution), kernel_size=1, padding='same', name='from_rgb_2')(inputs)
-        from_rgb_2 = d_block(from_rgb_2)
+        images_shape = (None,)+(int(2.0**self.current_resolution),)*self.dimensionality + (self.num_channels,)
+        alpha_shape = (1,)
 
-        lerp_input = self._weighted_sum()([from_rgb_1, from_rgb_2, alpha])
+        self.build([images_shape, alpha_shape])
 
-        output = self.growing_discriminator(lerp_input)
+    def call(self, inputs):
 
-        score_output = layers.Lambda(lambda x: x[...,0])(output)
-        label_output = layers.Lambda(lambda x: x[...,1:])(output)
+        images, alpha = inputs
 
-        self.growing_discriminator = tf.keras.Sequential([d_block, self.growing_discriminator])
-        self.train_discriminator = models.Model(inputs=[inputs, alpha], outputs=[score_output, label_output], name=self.name)
+        # To bring to the right number of filters
+        x = self.FadeConv(images)
+        y = self.highest_resolution_block(x)
+        x = self.discriminator_base(images, y, alpha)
 
-    def get_current_resolution(self):
-        return self.current_resolution
+        for d_block in self.resolution_blocks[::-1]:
+            x = d_block(x)
 
-    def get_trainable_discriminator(self):
-        return self.train_discriminator
-
-    def get_inference_discriminator(self):
-        raise NotImplementedError
+        return self.discriminator_head(x)
